use static ffmpeg over canonical ffmpeg.
This fixes #1 and #2. The static ffmpeg build lives at
https://github.com/heroku/sffmpeg. Since it's statically built it can
run across all heroku stacks. In addition, since we're controlling the
build we've cut a lot of the extra libs that aren't necessary. This has
cut the stack image impact from 200mb to 30mb. If we're missing
anything, we can evaluate that as the need arises.