Refactor bin/compile
- Remove cache of deb packages

Previously we stored a cache of the static ffmpeg binary, however both the cache and the binary are stored on S3 so this does not improve compile times. By moving to not use the cache and deleting old contents we can reduce cache size while preserving compile times

- Move location of `dpkg -x` to be logically consistent with installing poppler

Previously we were installing ffmpeg and poppler via the apt-get mechanism. With the move to have FFMPEG be a simple curl to disk, it makes more sense to have the `dpkg -x` logic be with the poppler download code.

- Add additional logging

Log download and install for poppler and ffmpegcache. Adds the version of ffmpeg to the log output.