Fix Dependabot config
Resolves:

```
The property '#/version' value 1 did not match one of the following values: 2
```

The PR limit has been removed for this repository, since it's unnecessary (it
only has a single test) and slows down trying to update dependencies when
we're very behind (eg when I'm trying to update to Hatchet 8.x now).

The explicit label has also been removed since the default labels are better.